+++ cmd/gateway/main.go
@@ -13,8 +13,7 @@ import (
 	"github.com/go-kratos/gateway/endpoint"
 	"github.com/go-kratos/gateway/endpoint/cors"
 	"github.com/go-kratos/gateway/endpoint/dyeing"
-	"github.com/go-kratos/gateway/endpoint/tracing"
-	"github.com/go-kratos/gateway/endpoint/tracing/otel"
+	"github.com/go-kratos/gateway/endpoint/otel"
 	"github.com/go-kratos/gateway/proxy"
 	"github.com/go-kratos/gateway/server"
 	"github.com/hashicorp/consul/api"
@@ -74,7 +73,7 @@ func middlewareFactory(c *configv1.Middleware) (endpoint.Middleware, error) {
 		return cors.Middleware(c)
 	case dyeing.Name:
 		return dyeing.Middleware(c)
-	case tracing.Otel:
+	case otel.Name:
 		return otel.Middleware(c)
 	default:
 		return nil, fmt.Errorf("not found middleware: %s", c.Name)
